# Conflicts:
#	src/main/java/com/example/CafeAutoSystem/review/controller/OrderController.java
#	src/main/java/com/example/CafeAutoSystem/review/entity/CafeOrder.java
#	src/main/java/com/example/CafeAutoSystem/review/entity/ReviewAnalysis.java

+++ src/main/java/com/example/CafeAutoSystem/common/entity/PurchaseOrderEntity.java
@@ -77,7 +77,7 @@ public class PurchaseOrderEntity extends BaseTime {
                 b.vendorName(this.vendorIngredient.getVendor().getVendorName());
             }
             if (this.vendorIngredient.getIngredient() != null) {
-                b.ingredientId(this.vendorIngredient.getIngredient().getIngredientId());
+                b.ingredientId(this.vendorIngredient.getIngredient().getIngredientId().intValue());
                 b.ingredientName(this.vendorIngredient.getIngredient().getIngredientName());
             }
         }
